data_collector/requirements.txt
requests
pandas
clickhouse-driver
PyYAML
schedule
python-dotenv
jdatetime